third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement.getBBox-03-expected.txt

This is a testharness.js-based test.
[FAIL] rect1
  assert_equals: rect1: x expected 0 but got 1
[FAIL] rect2
  assert_equals: rect2: x expected 0 but got 1
[FAIL] circle
  assert_equals: circle: x expected 0 but got 1
[FAIL] ellipse3
  assert_equals: ellipse3: x expected 0 but got 1
[FAIL] ellipse4
  assert_equals: ellipse4: x expected 0 but got 1
[FAIL] image3
  assert_equals: image3: x expected 0 but got 1
[FAIL] image4
  assert_equals: image4: x expected 0 but got 1
[FAIL] foreign1
  assert_equals: foreign1: x expected 0 but got 1
[FAIL] foreign2
  assert_equals: foreign2: x expected 0 but got 1
Harness: the test ran to completion.
